Refactor: Quantization (#454)
* split quant ops

* isolate the quantization modeling code

* properly set the _quant_C module

* import error fixes

* raise importerror only when the specific quant is called

* missed one

* roll back cc for hadamard

* address comments

* raise error in __init__ class

* missed it in gptq

* ruff
