+++ aphrodite/quantization/awq.py
@@ -1,16 +1,20 @@
 from typing import Any, Dict, List, Optional
+from contextlib import suppress
 
 import torch
 from torch.nn.parameter import Parameter
 
-from aphrodite._C import ops
 from aphrodite.modeling.layers.fused_moe import (moe_align_block_size,
                                                  fused_moe, fused_topk)
 from aphrodite.modeling.layers.linear import (LinearMethodBase,
                                               set_weight_attrs)
-from aphrodite.modeling.layers.quantization.base_config import (
+from aphrodite.quantization.base_config import (
     QuantizationConfig)
 
+HAS_QUANTS = False
+with suppress(ImportError):
+    from aphrodite._quant_C import quant_ops as ops
+    HAS_QUANTS = True
 
 class AWQConfig(QuantizationConfig):
     """Config class for AWQ.
@@ -24,6 +28,8 @@ class AWQConfig(QuantizationConfig):
         group_size: int,
         zero_point: bool,
     ) -> None:
+        if not HAS_QUANTS:
+            raise ImportError("Could not find the quantization kernels.")
         self.weight_bits = weight_bits
         self.group_size = group_size
         self.zero_point = zero_point
@@ -178,7 +184,7 @@ class AWQLinearMethod(LinearMethodBase):
             out = torch.matmul(reshaped_x, out)
         else:
             out = ops.awq_gemm(reshaped_x, qweight, scales, qzeros,
-                               pack_factor)
+                            pack_factor)
         if bias is not None:
             out = out + bias
         return out.reshape(out_shape)
