+++ aphrodite/quantization/gguf.py
@@ -1,13 +1,18 @@
 from typing import Any, Dict, List, Optional
+from contextlib import suppress
 
 import torch
 from torch.nn.parameter import Parameter
 
-from aphrodite._C import ops
 from aphrodite.modeling.layers.linear import (LinearMethodBase,
                                               set_weight_attrs)
-from aphrodite.modeling.layers.quantization.base_config import (
+from aphrodite.quantization.base_config import (
     QuantizationConfig)
+
+HAS_QUANTS = False
+with suppress(ImportError):
+    from aphrodite._quant_C import quant_ops as ops
+    HAS_QUANTS = True
 
 GGML_QUANT_SIZES = {
     0: (1, 4),  # F32
@@ -85,6 +90,8 @@ class GGUFLinearMethod(LinearMethodBase):
     """
 
     def __init__(self, quant_config: GGUFConfig):
+        if not HAS_QUANTS:
+            raise ImportError("Could not find the quantization kernels.")
         self.quant_config = quant_config
 
     def create_weights(self, input_size_per_partition: int,
@@ -127,13 +134,13 @@ class GGUFLinearMethod(LinearMethodBase):
         xshape = x.view(-1, x.shape[-1])
         if xshape.shape[0] == 1:
             out = ops.ggml_mul_mat_vec_a8(weight, reshaped_x, weight_type,
-                                          outfeatures)
+                                        outfeatures)
         elif xshape.shape[0] < 8 and weight_type < 16:
             out = ops.ggml_mul_mat_a8(weight, reshaped_x, weight_type,
-                                      outfeatures)
+                                    outfeatures)
         else:
             weight = ops.ggml_dequantize(weight, weight_type, outfeatures,
-                                         infeatures)
+                                        infeatures)
             out = reshaped_x @ weight.T
 
         if bias is not None:
@@ -156,7 +163,7 @@ class GGUFLinearMethod(LinearMethodBase):
                                    dim=0,
                                    index=x_flat)
         dequant = ops.ggml_dequantize(quant, weight_type, hidden_size,
-                                      x_flat.shape[0])
+                                    x_flat.shape[0])
         return dequant.view(*x.shape, hidden_size)
 
     def apply_moe_weights(self, w1: Dict[str,
