+++ kernels/pybind.cpp
@@ -59,30 +59,6 @@ PYBIND11_MODULE(TORCH_EXTENSION_NAME, m) {
     "batched_rotary_embedding",
     &batched_rotary_embedding,
     "Apply batched GPT-NeoX or GPT-J style rotary embedding to query and key");
-
-#ifndef USE_ROCM
-  // Quantization ops
-  ops.def("aqlm_gemm", &aqlm_gemm, "Quantized GEMM for AQLM");
-  ops.def("awq_gemm", &awq_gemm, "Quantized GEMM for AWQ");
-  ops.def("awq_dequantize", &awq_dequantize, "Dequantization for AWQ");
-  ops.def("awq_group_gemm", &awq_group_gemm, "Grouped Quantized GEMM for AWQ");
-  ops.def("autoquant_convert_s4_k_m8", &autoquant_convert_s4_k_m8, "convert kernel.");
-  ops.def("autoquant_s4_f16_gemm", &autoquant_s4_f16_gemm, "weight int4 activation float16 gemm kernel.");
-  ops.def("quip_decompress", &decompress_e8p_origorder, "decompress_packed_e8p");
-  ops.def("quip_gemv", &e8p_mm_origorder, "e8p_mm_origorder");
-  ops.def("marlin_gemm", &marlin_gemm, "Marlin Optimized Quantized GEMM for GPTQ");
-#endif
-  ops.def("gptq_gemm", &gptq_gemm, "Quantized GEMM for GPTQ");
-  ops.def("gptq_shuffle", &gptq_shuffle, "Post processing for GPTQ");
-  ops.def("group_gptq_gemm", &group_gptq_gemm, "Grouped Quantized GEMM for GPTQ");
-  ops.def("dequant_gptq", &dequant_gptq, "Dequantize gptq weight to half");
-  ops.def("squeezellm_gemm", &squeezellm_gemm, "Quantized GEMM for SqueezeLLM");
-  ops.def("ggml_dequantize", &ggml_dequantize, "ggml_dequantize");
-  ops.def("ggml_mul_mat_vec", &ggml_mul_mat_vec, "ggml_mul_mat_vec");
-  ops.def("ggml_mul_mat_vec_a8", &ggml_mul_mat_vec_a8, "ggml_mul_mat_vec_a8");
-  ops.def("ggml_mul_mat_a8", &ggml_mul_mat_a8, "ggml_mul_mat_a8");
-  ops.def("exl2_make_q_matrix",&make_q_matrix, "preprocess for exl2");
-  ops.def("exl2_gemm", &exl2_gemm, "exl2 gemm");
   
   ops.def("moe_align_block_size",
           &moe_align_block_size,
